Added Legacy Models we used for testing
This file is depreciated and it is a proof of work. It contains 2 models and some obsolete data processing methods, ConvNetwork and ConvNetwork_2. Most of their useful features was incorporated into CNN_Model.ipynb. 
When ConvNetwork was loaded with cnn_test_1.pth (which contains a set of weight), it produces around 75% testing accuracy.